[Packaging] Generate NuGet imports on installing NuGet.Build.Packaging
MSBuild imports are not generated when the NuGet.Build.Packaging
NuGet package is installed. These imports will import the MSBuild
targets for the NuGet.Build.Packaging NuGet package. The generated
imports are used by the NuGet.Packaging.Authoring MSBuild targets
that are included with the addin.

The NuGet.Build.Packaging NuGet package is now added to NuGet packaging
projects as a development dependency by setting PrivateAssets to All
for its PackageReference.